Fix overrides of AssetBundle.loadStructuredData().
The base class takes a callback that can return anything, but two
subclasses restrict that to callbacks that must return futures.

As far as I can tell, the restriction isn't needed. Both subclasses
already would handle callbacks that return immediate values (one is
in an async method and the other passes the result to Future.then()),
so I just widened the type in the override.