Make the Matcher.matches() overrides allow any type.
This is arguably a case where the covariant override is legitimate,
though I personally prefer the change I'm doing here.

Matcher.matches() takes Object for the object being matched. The
subclasses here were tightening that to only the type that they know
how to match. This means a match that fails because the expected value
is of the wrong type entirely will not produce a match failure, but
will instead generate a type error in the bowels of the matcher library.

That feels weird to me. I'd rather have any failure inside matching be
handled as a match failure. So, here, I widen the overridden methods to
accept any type. Then I manually test the type and cause the match to
fail if it's not the exected type.