This change is dubious, but I thought I would give it a try. It does fix the error and improve the type safety, at the expense of a little complexity.
My understanding is that this exposes two user-facing classes Markdown
and MarkdownRaw. Each has its own corresponding style class
MarkdownStyleRaw and MarkdownStyle.

Markdown overrides createMarkdownBody() and tightens the markdownStyle
to require its more specific MarkdownStyle class. This means that code
can't heterogeneously work with Markdown and MarkdownRaw unless it's
careful to make sure each gets the right style.

My guess was that from the user's perspective, these are really two
separate classes that happen to share a base class mainly for code
reuse in the implementation and not because they are really
substitutable.

So, what I did was introduce a generic base class (MarkdownBase) that
takes the style as a type parameter. The two user-facing classes,
Markdown and MarkdownRaw extend that and plug-in their preferred style
class.

This means you can't as easily write code that works with a mixture of
the two, but if that's not needed, this might be a workable solution.